+++ src/css/index.scss
@@ -11,6 +11,7 @@
 
 @import './app.scss';
 @import './views/register.scss';
+@import './views/login.scss';
 
 * {
   box-sizing: border-box;
