+++ src/css/views/login.scss
@@ -1,0 +1,77 @@
+.login {
+  display: flex;
+  min-height: 100vh;
+  align-items: center;
+  flex-wrap: wrap;
+
+  > div {
+    width: 50%;
+    min-width: 484px;
+    min-height: 100vh;
+    padding: 50px;
+  }
+
+  .login__left {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    img {
+      width: 100%;
+      max-width: 500px;
+      align-self: flex-start;
+      margin: 50px auto;
+    }
+
+    p {
+      text-align: center;
+      font-size: 36px;
+      font-family: 'Teko';
+      max-width: 500px;
+
+      span {
+        display: block;
+        text-align: right;
+        margin-top: 20px;
+      }
+    }
+  }
+
+  .login__right {
+    background-color: #333;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    > form {
+      background-color: #fff;
+      padding: 20px;
+      width: 100%;
+      max-width: 400px;
+      -webkit-box-shadow: 5px 5px 5px 0px rgba(0, 0, 0, 0.75);
+      -moz-box-shadow: 5px 5px 5px 0px rgba(0, 0, 0, 0.75);
+      box-shadow: 5px 5px 5px 0px rgba(0, 0, 0, 0.75);
+
+      .input-wrapper__label {
+        font-size: 18px;
+      }
+
+      button {
+        width: 100%;
+        margin-left: 0;
+        font-size: 24px;
+      }
+
+      a {
+        text-align: center;
+        display: block;
+        margin-top: 10px;
+      }
+    }
+  }
+
+  @media (max-width: 967px) {
+    > div {
+      width: 100%;
+    }
+  }
+}
